Implemented AOP Concepts in Spring Basics, fixed NoteRepository for notess-app

+++ notess-api/src/main/java/com/example/demo/repository/NotesRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.Note;
 
 @Repository
-public interface NotesRepository extends CrudRepository<Note, Integer>{
+public interface NotesRepository extends CrudRepository<Note, Long>{
 
 }
